refactor(login register): sSO buttons grouping and refactor
https://github.com/zesty-io/website/pull/1542#pullrequestreview-1245439403
